backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy==1.26.3
pillow==10.2.0
# tensorflow==2.15.0  # Remove this if using PyTorch or Scikit-Learn
# torch==2.1.2      # Uncomment this if using PyTorch
# scikit-learn      # Uncomment this if using Scikit-Learn